+++ jirust-cli/src/runners/jira_cmd_runners/version_cmd_runner.rs
@@ -71,9 +71,13 @@ impl VersionCmdRunner {
             cfg: config,
             resolution_value: serde_json::from_str(cfg_file.get_standard_resolution().as_str())
                 .unwrap_or(Value::Null),
-            resolution_comment: serde_json::from_str(jira_doc_std_field!(cfg_file
-                .get_standard_resolution_comment()
-                .as_str()))
+            resolution_comment: serde_json::from_str(
+                format!(
+                    "{{body: {}}}",
+                    jira_doc_std_field!(cfg_file.get_standard_resolution_comment().as_str())
+                )
+                .as_str(),
+            )
             .unwrap_or(Value::Null),
             resolution_transition_name: cfg_file.get_transition_name("resolve"),
         }