+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 ## [Unreleased]
+### Fixed
+- Fixed the format used in jira_doc_std_field macro to allow to use it properly
 
 ## [0.10.0] 2024-11-21
 ### Added